Create abstraction for PartitionOffsetFetcher and PartitionCountFetcher
