+++ pinot-controller/src/main/java/com/linkedin/pinot/controller/helix/core/PinotTableIdealStateBuilder.java
@@ -25,14 +25,10 @@ import com.linkedin.pinot.common.utils.CommonConstants.Helix;
 import com.linkedin.pinot.common.utils.StringUtil;
 import com.linkedin.pinot.common.utils.retry.RetryPolicies;
 import com.linkedin.pinot.controller.helix.core.realtime.PinotLLCRealtimeSegmentManager;
-import com.linkedin.pinot.core.realtime.stream.StreamConsumerFactory;
-import com.linkedin.pinot.core.realtime.stream.StreamConsumerFactoryProvider;
+import com.linkedin.pinot.core.realtime.stream.PartitionCountFetcher;
 import com.linkedin.pinot.core.realtime.stream.StreamMetadata;
-import com.linkedin.pinot.core.realtime.stream.StreamMetadataProvider;
-import com.linkedin.pinot.core.realtime.stream.TransientConsumerException;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Callable;
 import org.apache.helix.HelixAdmin;
 import org.apache.helix.ZNRecord;
 import org.apache.helix.model.IdealState;
@@ -142,14 +138,14 @@ public class PinotTableIdealStateBuilder {
     }
   }
 
-  public static int getPartitionCount(StreamMetadata kafkaMetadata) {
-    KafkaPartitionsCountFetcher fetcher = new KafkaPartitionsCountFetcher(kafkaMetadata);
+  public static int getPartitionCount(StreamMetadata streamMetadata) {
+    PartitionCountFetcher partitionCountFetcher = new PartitionCountFetcher(streamMetadata);
     try {
-      RetryPolicies.noDelayRetryPolicy(3).attempt(fetcher);
-      return fetcher.getPartitionCount();
+      RetryPolicies.noDelayRetryPolicy(3).attempt(partitionCountFetcher);
+      return partitionCountFetcher.getPartitionCount();
     } catch (Exception e) {
-      Exception fetcherException = fetcher.getException();
-      LOGGER.error("Could not get partition count for {}", kafkaMetadata.getKafkaTopicName(), fetcherException);
+      Exception fetcherException = partitionCountFetcher.getException();
+      LOGGER.error("Could not get partition count for {}", streamMetadata.getKafkaTopicName(), fetcherException);
       throw new RuntimeException(fetcherException);
     }
   }
@@ -204,56 +200,5 @@ public class PinotTableIdealStateBuilder {
       groupId = streamProviderConfig.get(keyOfGroupId);
     }
     return groupId;
-  }
-
-  // TODO: move this class to kafka specific package
-  private static class KafkaPartitionsCountFetcher implements Callable<Boolean> {
-    private int _partitionCount = -1;
-    private final StreamMetadata _streamMetadata;
-    private StreamConsumerFactory _streamConsumerFactory;
-    private Exception _exception;
-
-    private KafkaPartitionsCountFetcher(StreamMetadata streamMetadata) {
-      _streamMetadata = streamMetadata;
-      _streamConsumerFactory = StreamConsumerFactoryProvider.create(_streamMetadata);
-    }
-
-    private int getPartitionCount() {
-      return _partitionCount;
-    }
-
-    private Exception getException() {
-      return _exception;
-    }
-
-    @Override
-    public Boolean call() throws Exception {
-      final String bootstrapHosts = _streamMetadata.getBootstrapHosts();
-      final String kafkaTopicName = _streamMetadata.getKafkaTopicName();
-      if (bootstrapHosts == null || bootstrapHosts.isEmpty()) {
-        LOGGER.warn("Could not get partition count for topic {}. Invalid config for bootstrap hosts:'{}'",
-            kafkaTopicName, bootstrapHosts);
-        throw new RuntimeException(
-            "Invalid value for " + Helix.DataSource.Realtime.Kafka.KAFKA_BROKER_LIST + ":'" + bootstrapHosts + "'");
-      }
-
-      String clientId = PinotTableIdealStateBuilder.class.getSimpleName() + "-" + kafkaTopicName;
-      try (StreamMetadataProvider streamMetadataProvider = _streamConsumerFactory.createStreamMetadataProvider(clientId)) {
-        _partitionCount = streamMetadataProvider.fetchPartitionCount(/*maxWaitTimeMs=*/5000L);
-        if (_exception != null) {
-          // We had at least one failure, but succeeded now. Log an info
-          LOGGER.info("Successfully retrieved partition count as {} for topic {}", _partitionCount, kafkaTopicName);
-        }
-        return Boolean.TRUE;
-      } catch (TransientConsumerException e) {
-        LOGGER.warn("Could not get partition count for topic {}", kafkaTopicName, e);
-        _exception = e;
-        return Boolean.FALSE;
-      } catch (Exception e) {
-        LOGGER.warn("Could not get partition count for topic {}", kafkaTopicName, e);
-        _exception = e;
-        throw e;
-      }
-    }
   }
 }
