+++ pinot-core/src/main/java/com/linkedin/pinot/core/realtime/stream/PartitionOffsetFetcher.java
@@ -1,0 +1,84 @@
+/**
+ * Copyright (C) 2014-2018 LinkedIn Corp. (pinot-core@linkedin.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.linkedin.pinot.core.realtime.stream;
+
+import java.util.concurrent.Callable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+
+/**
+ * Fetches the partition offset for a stream given the offset criteria, using the {@link StreamMetadataProvider}
+ */
+public class PartitionOffsetFetcher implements Callable<Boolean> {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(PartitionOffsetFetcher.class);
+  private static final int STREAM_PARTITION_OFFSET_FETCH_TIMEOUT_MILLIS = 10000;
+
+  private final String _topicName;
+  private final String _offsetCriteria;
+  private final int _partitionId;
+
+  private Exception _exception = null;
+  private long _offset = -1;
+  private StreamConsumerFactory _streamConsumerFactory;
+  StreamMetadata _streamMetadata;
+
+  public PartitionOffsetFetcher(final String offsetCriteria, int partitionId, StreamMetadata streamMetadata) {
+    _offsetCriteria = offsetCriteria;
+    _partitionId = partitionId;
+    _streamMetadata = streamMetadata;
+    _streamConsumerFactory = StreamConsumerFactoryProvider.create(streamMetadata);
+    _topicName = streamMetadata.getKafkaTopicName();
+  }
+
+  public long getOffset() {
+    return _offset;
+  }
+
+  public Exception getException() {
+    return _exception;
+  }
+
+  /**
+   * Callable to fetch the offset of the partition given the stream metadata and offset criteria
+   * @return
+   * @throws Exception
+   */
+  @Override
+  public Boolean call() throws Exception {
+    String clientId = PartitionOffsetFetcher.class.getSimpleName() + "-" + _topicName + "-" + _partitionId;
+    try (
+        StreamMetadataProvider streamMetadataProvider = _streamConsumerFactory.createPartitionMetadataProvider(clientId,
+            _partitionId)) {
+      _offset =
+          streamMetadataProvider.fetchPartitionOffset(_offsetCriteria, STREAM_PARTITION_OFFSET_FETCH_TIMEOUT_MILLIS);
+      if (_exception != null) {
+        LOGGER.info("Successfully retrieved offset({}) for stream topic {} partition {}", _offset, _topicName,
+            _partitionId);
+      }
+      return Boolean.TRUE;
+    } catch (TransientConsumerException e) {
+      LOGGER.warn("Temporary exception when fetching offset for topic {} partition {}:{}", _topicName, _partitionId,
+          e.getMessage());
+      _exception = e;
+      return Boolean.FALSE;
+    } catch (Exception e) {
+      _exception = e;
+      throw e;
+    }
+  }
+}
